requirements.txt
alibabacloud-alidns20150109==3.5.0
tencentcloud-sdk-python==3.0.806
python-dateutil==2.9.0.post0
#prettytable==3.10.0
requests==2.28.1